Initial Consolidation of bridge
